Add JavaScript solution to find which kid can have the greatest number of candies after extra candies are given
This repository contains a JavaScript solution for a coding challenge that involves finding which kid can have the greatest number of candies after extra candies are given. The solution is implemented in a function called 'kidsWithCandies', which takes an array of integers representing the number of candies each kid has and an extra number of candies as input. The function returns a boolean array of the same length as the input array, indicating whether each kid can have the greatest number of candies after the extra candies are given.

The 'kidsWithCandies' function works by finding the maximum number of candies in the input array, then iterating over the array to check if adding the extra candies to each kid's candies will result in them having the greatest number of candies. The function returns an array of boolean values indicating the result for each kid.

This solution can be used as a reference for other coding challenges involving array manipulation and conditional statements in JavaScript.