Challenge 2: Counter Function
The Counter Function challenge is about creating a function that takes an integer n and returns another function that acts as a counter.

The returned function should initially return n and then return 1 more than the previous value every subsequent time it is called. For example, if n is 10, then the first call to the returned function should return 10, the second call should return 11, and so on.